test: working test for removing access
